[yarpRTMPose] Dir structure refactored. Added integration test
