+++ yarpRTMPose/Dockerfile
@@ -111,12 +111,10 @@ RUN cd /robotology && git clone https://github.com/robotology/icub-main.git \
 RUN cd /robotology && git clone https://github.com/robotology/icub-contrib-common.git \
 	&& cd icub-contrib-common && mkdir build && cd build && cmake .. && make && make install
 
-ENV PYTHONPATH="/robotology/yarp/build/lib/python3:/root/yarpPose"
+ENV PYTHONPATH="/robotology/yarp/build/lib/python3:/yarpRTMPose/src/python"
 # We need tools and resources from mmdeploy 
 RUN git clone -b main https://github.com/open-mmlab/mmdeploy.git
 
 # Install the application
 COPY . /yarpRTMPose
 RUN /yarpRTMPose/app/scripts/deploy.sh
-
-ENTRYPOINT [ "/bin/bash", "/yarpRTMPose/docker/entrypoint.sh" ]
