+++ yarpRTMPose/app/conf/grabber.ini
@@ -1,2 +1,2 @@
 device ffmpeg_grabber
-source /home/fbrand/Videos/trial_videos/MeWalking.avi 
+source /Videos/trial_videos/MeWalking.avi
