+++ yarpRTMPose/app/conf/yarpRTMPose.ini
@@ -1,6 +1,6 @@
 # These must correspond to the input image height and width
-height 640
-width 480
+height 720
+width 1280
 
 # Please use the following models for the time being
 det_model_path /mmdeploy/rtmpose-ort/rtmdet-nano
