--- yarpRTMPose/docker/entrypoint.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-yarp detect --write
-yarprun --server /container-rtmpose --log &
-
-/bin/bash
