+++ yarpRTMPose/docker/start_docker.sh
@@ -1,16 +1,30 @@
-if [[ -z $1 ]]
-then
-  DATA_DIR_PATH=~/Videos
-else
-  DATA_DIR_PATH=$1
-fi
+#!/bin/bash
 
-if [[ -z $2 ]]
-then
-  IMG=fbrand-new/yarprtmpose:devel_cudnn
-else
-  IMG=$2
-fi
+DATA_DIR_PATH=~/Videos
+IMG=fbrand-new/yarprtmpose:devel_cudnn
+CMD=bash
+
+while getopts ":i:d:c:" opt; do
+  case $opt in
+    i)  
+      IMG="$OPTARG"
+    ;;
+    d) DATA_DIR_PATH="$OPTARG"
+    ;;
+    c) 
+      CMD="$OPTARG"
+    ;;
+    \?) echo "Invalid option -$OPTARG" >&3
+    exit 1
+    ;;
+  esac
+
+  case $OPTARG in
+    -*) echo "Option $opt needs a valid argument"
+    exit 1
+    ;;
+  esac
+done
 
 DATA_DIR=$(basename ${DATA_DIR_PATH})
 
@@ -18,4 +32,5 @@ docker run -it --rm --privileged --gpus=all --network host --pid host\
   -v /etc/localtime:/etc/localtime -v /tmp/.X11-unix:/tmp/.X11-unix \
   -v ${XAUTHORITY}:/root/.Xauthority -e DISPLAY=$DISPLAY \
   -e QT_X11_NO_MITSHM=1 -v /etc/hosts:/etc/hosts -v $DATA_DIR_PATH:/${DATA_DIR} \
-  $IMG bash
+  $IMG $CMD
+
